Refactor: Separate JavaScript logic into modules
I've separated your JavaScript codebase into distinct modules for better organization and extensibility:

- `js/data.js`: Contains the `notes` array with note metadata.
- `js/ui.js`: Handles DOM manipulation, including rendering the note list, note content, table of contents, and navigation buttons.
- `js/main.js`: Acts as the application coordinator, handling event listeners, routing, and orchestrating calls to the data and UI modules.

I also updated `index.html` to load these JavaScript files as ES6 modules in the correct order. This refactoring improves code clarity, maintainability, and makes it easier to add new features or modify existing ones.
