+++ index.html
@@ -27,6 +27,8 @@
   </div>
   <script src="lib/marked.min.js"></script>
   <script src="js/toc.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/ui.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
